v1.0.0 - HA Core Compliance Release
