--- .vscode/launch.json
@@ -1,28 +1,0 @@
-{
-    "configurations": [
-        {
-            "justMyCode": false,
-            "name": "Python: Attach Local",
-            "pathMappings": [
-                {
-                    "localRoot": "${workspaceFolder}",
-                    "remoteRoot": "."
-                }
-            ],
-            "request": "attach",
-            "type": "debugpy"
-        },
-        {
-            "name": "Python: Attach Remote",
-            "pathMappings": [
-                {
-                    "localRoot": "${workspaceFolder}",
-                    "remoteRoot": "/usr/src/homeassistant"
-                }
-            ],
-            "request": "attach",
-            "type": "debugpy"
-        }
-    ],
-    "version": "0.2.0"
-}
