+++ custom_components/garmin_connect/manifest.json
@@ -1,12 +1,16 @@
 {
   "domain": "garmin_connect",
   "name": "Garmin Connect",
-  "codeowners": ["@cyberjunky"],
+  "codeowners": [
+    "@cyberjunky"
+  ],
   "config_flow": true,
   "dependencies": [],
   "documentation": "https://github.com/cyberjunky/home-assistant-garmin_connect",
   "iot_class": "cloud_polling",
   "issue_tracker": "https://github.com/cyberjunky/home-assistant-garmin_connect/issues",
-  "requirements": ["garminconnect>=0.2.33"],
-  "version": "0.2.39-beta-01"
+  "requirements": [
+    "garminconnect>=0.2.37"
+  ],
+  "version": "1.0.0-beta-01"
 }
