+++ custom_components/garmin_connect/strings.json
@@ -33,16 +33,5 @@
       "already_configured": "[%key:common::config_flow::abort::already_configured_account%]",
       "reauth_successful": "[%key:common::config_flow::abort::reauth_successful%]"
     }
-  },
-  "options": {
-    "step": {
-      "init": {
-        "title": "Configure Sensor Groups",
-        "description": "Select which sensor groups to enable. Individual sensors within enabled groups can still be disabled in the entity settings. Changes will be applied after reloading the integration.",
-        "data": {
-          "sensor_groups": "Enabled Sensor Groups"
-        }
-      }
-    }
   }
 }
