+++ hacs.json
@@ -1,6 +1,6 @@
 {
     "name": "Garmin Connect",
-    "render_readme": true,
-    "homeassistant": "2025.1.4",
-    "hacs": "1.34.0"
+    "homeassistant": "2025.2.4",
+    "hacs": "2.0.5",
+    "render_readme": true
 }
