+++ pyproject.toml
@@ -1,119 +1,5 @@
 [tool.isort]
-# https://github.com/PyCQA/isort/wiki/isort-Settings
 profile = "black"
-# will group `import x` and `from x import` of the same module.
 force_sort_within_sections = true
 known_first_party = ["custom_components"]
 combine_as_imports = true
-
-[tool.pytest.ini_options]
-asyncio_mode = "auto"
-addopts = "-rxf -l --cov=./ --cov-report=xml"
-filterwarnings = [
-    "ignore::DeprecationWarning",
-    "ignore:It is recommended to use web.AppKey instances for keys",
-]
-log_format = "%(asctime)s.%(msecs)03d %(levelname)-8s %(threadName)s %(name)s:%(filename)s:%(lineno)s %(message)s"
-log_date_format = "%Y-%m-%d %H:%M:%S"
-
-norecursedirs = [".git"]
-
-[tool.ruff]
-fix = true
-line-length = 100
-show-fixes = true
-target-version = "py312"
-
-[tool.lint]
-select = ["ALL"]
-ignore = [
-    "ANN001",
-    "ANN002",
-    "ANN003",
-    "ANN101",
-    "ANN201",
-    "ANN202",
-    "ANN204",
-    "ANN205",
-    "ANN401",
-    "ARG001",
-    "ARG002",
-    "ARG005",
-    "ASYNC110",
-    "BLE001",
-    "C901",
-    "COM812",
-    "D100",
-    "D101",
-    "D102",
-    "D103",
-    "D105",
-    "D107",
-    "D202",
-    "D203",
-    "D205",
-    "D213",
-    "D400",
-    "D401",
-    "D415",
-    "E501",
-    "E713",
-    "EM101",
-    "EM102",
-    "F401",
-    "FBT001",
-    "FBT002",
-    "FBT003",
-    "I001",
-    "INP001",
-    "ISC001",
-    "N812",
-    "N818",
-    "PERF401",
-    "PERF402",
-    "PIE804",
-    "PLR0912",
-    "PLR0913",
-    "PLR0915",
-    "PLR1714",
-    "PLR1722",
-    "PLR2004",
-    "PLR5501",
-    "PTH103",
-    "PTH107",
-    "PTH110",
-    "PTH113",
-    "PTH123",
-    "RET502",
-    "RET503",
-    "RET504",
-    "RET507",
-    "RSE102",
-    "RUF005",
-    "RUF012",
-    "RUF013",
-    "S105",
-    "S110",
-    "SIM102",
-    "SIM103",
-    "SIM105",
-    "SIM108",
-    "SIM110",
-    "SIM114",
-    "SIM117",
-    "SLF001",
-    "TCH001",
-    "TCH002",
-    "TCH003",
-    "TID252",
-    "TRY003",
-    "TRY201",
-    "TRY300",
-    "TRY301",
-    "TRY400",
-    "TRY401",
-    "UP040",
-]
-
-fixable = ["ALL"]
-unfixable = []
