--- requirements_core_min.txt
@@ -1,1 +1,0 @@
-homeassistant==2025.1.4
