+++ requirements_lint.txt
@@ -1,8 +1,9 @@
---requirement requirements_base.txt
 codespell==2.4.1
-isort==7.0.0
-pre-commit==4.3.0
+mypy==1.19.1
+pre-commit==4.5.1
 pre-commit-hooks==6.0.0
-pyupgrade==3.21.0
-ruff==0.14.3
+pylint==4.0.4
+pyupgrade==3.21.2
+ruff==0.14.10
+types-pytz==2025.2.0.20251108
 vulture==2.14
