+++ scripts/install/core
@@ -4,4 +4,4 @@ set -e
 
 cd "$(dirname "$0")/../.."
 
-bash scripts/install/pip_packages --requirement requirements_core_min.txt
+bash scripts/install/pip_packages --requirement requirements.txt
