+++ scripts/install/pip_packages
@@ -6,5 +6,4 @@ python3 -m pip \
     install \
     --upgrade \
     --disable-pip-version-check \
-    --constraint constraints.txt \
     "${@}"
